Create README.md
Explaining it the best way I can to help you understand you just need the object data loaded or queued prior to using this otherwise you are waiting for results of the query itself. Meaning this algo is already finished at the moment the query is finished, the query you made for results was the only time consuming part about your data. Example generating 500,000 object arrays will take time to computer but to query 500,000 lines in advanced sorted can be reduced into ms.