Create nightmare.tsp.js
Okay okay, maybe it's complex and I figured out I did not solve the problem but a better solution to it. Well, This is the nightmare version to that code. Nightmare 0.0.1 version called salesmans nightmare machine computes shortest path faster than previously known. This won't be my last version but until my brain farts are in the fiat I will not fart on it much more.