Create water-tsp.js
Attempted to scale up the data to handle the ocean but it slows the code down compared to lightning version. I believe this is the lowest ms you can get fo query given my knowledge and current understanding of what needs to happen to skip a few more processes for known results. - It's much like brute force but given if you are unable to compute a problem passing it is going to always seem like it. 

Lightning Execution Time: 0.40 ms
Water Execution Time: 1.20 ms
Both have variations in time depending on previous data provided.