Update README.md
Added a screenshot of latest results. Here's the latest version, soon to be refined to handle 1 million objects in under 60 seconds. For now, this is all you get from me. Many claim to be smarter and wiser here's your chance to shine. Beat this, and I'll reveal what I've seen before you.