Create zap.py
First Starting Point: The first city in the sorted path. {'name': 'City229', 'x': 500, 'y': 125}
Second-to-Last Starting Point: The city that was second to last in the sorted array before adding the first city again. {'name': 'City732', 'x': 899, 'y': 169}
Last Starting Point: The first city that was appended to the end of the sorted array to form a closed loop. {'name': 'City229', 'x': 500, 'y': 125}

Sorting Time: This is the time taken to sort the array of cities. In this example, it took 0 seconds and 045 milliseconds. Sorting Time: 0s 045ms
After Sorted Adjusted Last to First: This is the time taken to add the first city to the end of the sorted array, ensuring the path forms a loop. This step also took 0 seconds and 045 milliseconds. After Sorted Adjusted Last to First: 0s 045ms (Note "To be clear this is 0 ms it uses end time - start time")