Create zap.py
To change arrays edit 
result = zap(cities)  
change cities to usa_states but the process is calculated and the drawing is purely for visual reference to what has been calculated. 

This is also created in the same instance with functions imports and timing calculations which aren't going to help you with speed but rather slow you down a bit.  Preloading this python then access it as a function will greatly speed up this process. Goodluck to all whom it serves a purpose for.